+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Bug fixes and small improvements:
 
 - Print calls and decision statistics in summary only if values are gathered. (:issue:`749`)
 - Log the thread name if :option:`-j` is used. (:issue:`752`)
+- Collapse also root directory if needed in nested HTML report. (:issue:`750`)
 
 Documentation: